+++ rector.php
@@ -3,6 +3,8 @@
 use Rector\CodeQuality\Rector\Concat\JoinStringConcatRector;
 use Rector\Config\RectorConfig;
 use Rector\PHPUnit\CodeQuality\Rector\Class_\PreferPHPUnitSelfCallRector;
+use Rector\PHPUnit\CodeQuality\Rector\Class_\PreferPHPUnitThisCallRector;
+use Rector\PHPUnit\Set\PHPUnitSetList;
 use Rector\Set\ValueObject\SetList;
 
 return RectorConfig::configure()
@@ -10,10 +12,20 @@ return RectorConfig::configure()
         SetList::CODE_QUALITY,
         SetList::TYPE_DECLARATION,
         SetList::RECTOR_PRESET,
+        PHPUnitSetList::PHPUNIT_40,
+        PHPUnitSetList::PHPUNIT_50,
+        PHPUnitSetList::PHPUNIT_60,
+        PHPUnitSetList::PHPUNIT_70,
+        PHPUnitSetList::PHPUNIT_80,
+        PHPUnitSetList::PHPUNIT_90,
+        PHPUnitSetList::PHPUNIT_100,
+        PHPUnitSetList::PHPUNIT_110,
+        PHPUnitSetList::PHPUNIT_CODE_QUALITY,
     ])
     ->withPhpSets()
     ->withRules([PreferPHPUnitSelfCallRector::class])
     ->withSkip([
+        PreferPHPUnitThisCallRector::class, // breaks tests
         JoinStringConcatRector::class => [
             __DIR__ . '/tests/CSVArrayTest.php', // keep `\r\n` for readability
         ],