+++ tests/DnaSequenceTest.php
@@ -10,19 +10,19 @@ final class DnaSequenceTest extends TestCase
     public function testReverse(): void
     {
         $dnaSequence = new DnaSequence('ATGC');
-        self::assertEquals('CGTA', $dnaSequence->reverse());
+        self::assertSame('CGTA', $dnaSequence->reverse());
     }
 
     public function testComplement(): void
     {
         $dnaSequence = new DnaSequence('ATGC');
-        self::assertEquals('TACG', $dnaSequence->complement());
+        self::assertSame('TACG', $dnaSequence->complement());
     }
 
     public function testReverseComplement(): void
     {
         $dnaSequence = new DnaSequence('ATGC');
-        self::assertEquals('GCAT', $dnaSequence->reverseComplement());
+        self::assertSame('GCAT', $dnaSequence->reverseComplement());
     }
 
     public function testInvalidSequenceThrowsException(): void
@@ -34,8 +34,8 @@ final class DnaSequenceTest extends TestCase
     public function testEmptySequence(): void
     {
         $dnaSequence = new DnaSequence('');
-        self::assertEquals('', $dnaSequence->reverse());
-        self::assertEquals('', $dnaSequence->complement());
-        self::assertEquals('', $dnaSequence->reverseComplement());
+        self::assertSame('', $dnaSequence->reverse());
+        self::assertSame('', $dnaSequence->complement());
+        self::assertSame('', $dnaSequence->reverseComplement());
     }
 }